# openapi.yaml
openapi: 3.0.3
info:
  title: FerroS3 API
  version: 0.1.0
  description: S3-compatible filesystem-backed proxy for local directories.
servers:
  - url: http://127.0.0.1:8080
    description: Local default
...
paths:
...
  /_admin/presign:
    post:
      tags: [Admin]
      summary: Generate a presigned URL
      description: Generates a presigned URL using the configured access key and secret key.
      parameters:
        - name: bucket
          in: query
          required: true
          schema:
            type: string
        - name: key
          in: query
          required: true
          schema:
            type: string
        - name: method
          in: query
          required: true
          schema:
            type: string
            enum: [GET, PUT, HEAD, DELETE]
        - name: expires
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
          description: Expiration time in seconds.
      responses:
        '200':
          description: Presigned URL
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Auth is not configured
          content:
            text/plain:
              schema:
                type: string
              example: Auth not configured
...